Corrected some instances in Exit where floats/doubled that were converted to ints where still being passed as floats/doubles in the getters and setters.
Made similar changes to Hole, also removing the canvas-based getters as they were no longer needed with dimensions ties to the canvas instead of percentages.

Made the same changes to Segment, also removing several of the addTail/s methods, modifying addTail/sLeft to accommodate fixation on the canvas size.

Made the same changes to Turn as mentioned above.

Modified Game Thread to get the canvas width and pass it into Game's new initializeBoard method before starting the game loop.

Modified Game heavily.  Retooled all the dimensions to actual values based on the canvas width passed in.  This obviates the need to calculate it at draw time, and simplified getting the positioning right without having to contend with too many variables such as float/double tolerances, off-by-one errors, GC/Thread stutter, and etc.  initializeBoard must be called before drawing and animating can work now, as the on-screen elements are now tied to the screen size.  Changed logic in animateThroughTurns to handle tails differently than heads.  Because there is still some potential for segments to wander from calculation inconsistencies, tails now position themselves based on the cell-size distance from their head and the turn, rather than the turn and their current position adjusted with velocity.

+++ app/src/main/java/edu/fullsail/whackapede/models/Exit.java
@@ -12,8 +12,8 @@ import androidx.annotation.Nullable;
 import java.util.ArrayList;
 
 class Exit {
-    private final float directionX;
-    private final float directionY;
+    private final int directionX;
+    private final int directionY;
 
     static final Exit exitTop    = new Exit(  0, -1 );
     static final Exit exitBottom = new Exit(  0,  1 );
@@ -56,7 +56,7 @@ class Exit {
         add( exitTop ); add( exitLeft );
     } };
 
-    private Exit( float directionX, float directionY ) {
+    private Exit( int directionX, int directionY ) {
         this.directionX = directionX;
         this.directionY = directionY;
     }
@@ -71,15 +71,11 @@ class Exit {
         return exit.getDirectionX() == directionX && exit.getDirectionY() == directionY;
     }
 
-    float getDirectionX() {
-        return directionX;
-    }
+    int getDirectionX() { return directionX; }
 
-    float getDirectionY() {
-        return directionY;
-    }
+    int getDirectionY() { return directionY; }
 
-    static Exit getExitReverseOf( float directionX, float directionY ) {
+    static Exit getExitReverseOf( int directionX, int directionY ) {
         return new Exit( 0 - directionX, 0 - directionY);
     }
 }
